test(refactor): 컴포넌트별 테스트 코드 분리
